Fix test driver, and other random changes
I managed to run into a double-free bug with it, but it's not
reproducible by running the test driver on only the file that ran into
the bug. Testing with ASan to confirm the origin of the bug is a job
for tomorrow.

(I'm running the software on the corpus of all files with .lzhal
extensions that are dumped into the src/bin directory in ebsrc. The
command I run is this:

find <directory-path> -name "*.lzhal" | xargs ./test

It is *really* really slow, mind. I'm not running any of that again
without first testing if ASan detects the bug quickly. (Or valgrind if
I can figure out how it works and it's not too slow. But ASan's there.)