Fix condition for skipping back reference candidates
ref_search assumes that the candidate it's looking for has a size of 4
or greater (since anything smaller would cause the back reference
command to take up more space than the literal bytes). The condition for
guarding against this in pack_optimal was typo'd to check the size of
the entire file, rather than the amount of space left in the file.

Fixes #8