Fix remaining bugs that ASan complains about
The first bug was a latent bug that was already there: rle_check always
checks for 16-bit RLE, including reading the two bytes for the potential
run, even if there's not two bytes available in the file to read. The
cleanest way to prevent that seemed to guard that entire section behind
a size check.

The other bugs were memory leaks that I introduced when handling malloc
failures.
In pack_context_alloc, if a malloc in uthash failed, it would leak the
element that was about to be inserted, because I forgot to free that
element before calling longjmp. This was fixed by actually freeing that
element.
In exhal_pack2, if pack_optimal failed, ctx would be leaked because I
immediately returned 0 before freeing the pack context. This was fixed
by letting a later point in the function handle return 0 instead.